backend/requirements.txt
# ============================================
# HealthMate Advanced Backend Requirements
# ============================================

fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
# sentence-transformers==2.2.2
# faiss-cpu==1.13.2
# logging==0.4.9.6
pydantic==2.5.0


# Web Framework
# fastapi==0.104.1
# uvicorn==0.24.0

# Data Validation
# pydantic==2.5.0

# HTTP Requests (for Grok + Gemini APIs)
# requests==2.31.0

# Environment Variables
# python-dotenv==1.0.0

# AI/ML - Embeddings & Vector Search (kept for RAG fallback)
# sentence-transformers==2.2.2

# faiss-cpu==1.13.2
# Numerical Computing
# numpy==1.24.3

# Logging
# logging==0.4.9.6

# Optional: For development
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0

